(Feat) DAO: Atualizado os métodos de DAO (Insert funcionando)

+++ Sistema.Model/DAO/DAO.cs
@@ -4,8 +4,11 @@ using Sistema.Model.Interfaces.IDAO;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.Common;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Reflection;
+using System.Threading.Tasks;
 
 public abstract class DAO<T> : IDAO<T> where T : IEntidade
 {
@@ -21,34 +24,50 @@ public abstract class DAO<T> : IDAO<T> where T : IEntidade
 
     public List<T> GetAll()
     {
-        List<T> result = new List<T>();
         try
         {
-            using (SqlConnection connection = ConnectionManager.GetConnection())
+            Type tipo = typeof(T);
+
+            string query = $"SELECT * FROM {TableName}";
+
+            List<T> result = new List<T>();
+
+            using (SqlCommand command = new SqlCommand(query, ConnectionManager.GetConnection()))
             {
-                string query = $"SELECT * FROM {TableName}";
-                using (SqlCommand command = new SqlCommand(query, connection))
+                ConnectionManager.OpenConnection();
+
+                using (SqlDataReader reader = command.ExecuteReader())
                 {
-                    using (SqlDataReader reader = command.ExecuteReader())
+                    while (reader.Read())
                     {
-                        while (reader.Read())
+                        T item = (T)Activator.CreateInstance(tipo);
+
+                        foreach (PropertyInfo property in tipo.GetProperties())
                         {
-                            T item = MapData(reader);
-                            result.Add(item);
+                            if (property.Name != "Id")
+                            {
+                                property.SetValue(item, reader[property.Name]);
+                            }
                         }
+
+                        result.Add(item);
                     }
                 }
+
+                ConnectionManager.CloseConnection();
             }
+
             return result;
-        }catch (SqlException sqlEx)
+        }
+        catch (SqlException sqlEx)
         {
             Console.WriteLine("Erro de banco: " + sqlEx.Message);
-            return result;
+            return null;
         }
         catch (Exception ex)
         {
-            Console.WriteLine("Erro: " +  ex.Message);
-            return result;
+            Console.WriteLine("Erro: " + ex.Message);
+            return null;
         }
     }
 
@@ -56,30 +75,50 @@ public abstract class DAO<T> : IDAO<T> where T : IEntidade
     {
         try
         {
-            using (SqlConnection connection = ConnectionManager.GetConnection())
+            Type tipo = typeof(T);
+
+            string query = $"SELECT * FROM {TableName} WHERE Id = @Id";
+
+            using (SqlCommand command = new SqlCommand(query, ConnectionManager.GetConnection()))
             {
-                string query = $"SELECT * FROM {TableName} WHERE Id = @Id";
-                using (SqlCommand command = new SqlCommand(query, connection))
+                command.Parameters.AddWithValue("@Id", id);
+
+                ConnectionManager.OpenConnection();
+
+                using (SqlDataReader reader = command.ExecuteReader())
                 {
-                    command.Parameters.AddWithValue("@Id", id);
-                    using (SqlDataReader reader = command.ExecuteReader())
+                    if (reader.Read())
                     {
-                        if (reader.Read())
+                        T item = (T)Activator.CreateInstance(tipo);
+
+                        foreach (PropertyInfo property in tipo.GetProperties())
                         {
-                            return MapData(reader);
+                            if (property.Name != "Id")
+                            {
+                                property.SetValue(item, reader[property.Name]);
+                            }
                         }
+
+                        ConnectionManager.CloseConnection();
+
+                        return item;
                     }
                 }
+
+                ConnectionManager.CloseConnection();
             }
-            return default;
-        }catch(SqlException sqlEx)
+
+            return default(T);
+        }
+        catch (SqlException sqlEx)
         {
             Console.WriteLine("Erro de banco: " + sqlEx.Message);
-            return default;
-        }catch(Exception ex)
+            return default(T);
+        }
+        catch (Exception ex)
         {
             Console.WriteLine("Erro: " + ex.Message);
-            return default;
+            return default(T);
         }
     }
 
@@ -87,84 +126,86 @@ public abstract class DAO<T> : IDAO<T> where T : IEntidade
     {
         try
         {
-            using (SqlConnection connection = ConnectionManager.GetConnection())
+            Type tipo = typeof(T);
+
+            PropertyInfo[] propriedades = tipo.GetProperties().Where(p => p.Name != "Id").ToArray();
+
+            string colunas = string.Join(", ", propriedades.Select(p => p.Name));
+            string parametros = string.Join(", ", propriedades.Select((p, i) => $"@p{i}"));
+
+            string query = $"INSERT INTO {TableName} ({colunas}) VALUES ({parametros})";
+
+            using (SqlCommand command = new SqlCommand(query, ConnectionManager.GetConnection()))
             {
-                // Obtém os nomes das colunas da tabela, excluindo a coluna 'Id'
-                var columnNames = typeof(T).GetProperties()
-                    .Where(prop => prop.Name != "Id")
-                    .Select(prop => prop.Name)
-                    .ToArray();
-
-                // Monta a lista de parâmetros da consulta SQL
-                var parameters = columnNames.Select(colName => $"@{colName}").ToArray();
-
-                // Monta a consulta SQL de inserção dinâmica
-                string query = $"INSERT INTO {TableName} ({string.Join(", ", columnNames)}) " +
-                               $"VALUES ({string.Join(", ", parameters)})";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
+                for (int i = 0; i < propriedades.Length; i++)
                 {
-                    // Configura os parâmetros da consulta SQL com os valores do objeto 'item'
-                    foreach (var propInfo in typeof(T).GetProperties())
-                    {
-                        if (propInfo.Name != "Id")
-                        {
-                            var paramValue = propInfo.GetValue(item);
-                            command.Parameters.AddWithValue($"@{propInfo.Name}", paramValue ?? DBNull.Value);
-                        }
-                    }
-
-                    command.ExecuteNonQuery();
+                    command.Parameters.AddWithValue($"@p{i}", propriedades[i].GetValue(item));
                 }
+
+                ConnectionManager.OpenConnection();
+                int count = command.ExecuteNonQuery();
+                ConnectionManager.CloseConnection();
+
+                return count > 0;
             }
-            return true;
-        }catch(SqlException sqlEx)
+        }
+        catch (SqlException sqlEx)
         {
             Console.WriteLine("Erro de banco: " + sqlEx.Message);
             return false;
-        }catch(Exception ex)
+        }
+        catch (Exception ex)
         {
             Console.WriteLine("Erro: " + ex.Message);
             return false;
         }
     }
 
+
     public bool Update(T item)
     {
         try
         {
-            using (SqlConnection connection = ConnectionManager.GetConnection())
+            Type tipo = typeof(T);
+
+            int id = (int)tipo.GetProperty("Id").GetValue(item);
+
+            if (id <= 0)
             {
-                // Obtém os nomes das colunas da tabela, excluindo a coluna 'Id'
-                var columnNames = typeof(T).GetProperties()
-                    .Where(prop => prop.Name != "Id")
-                    .Select(prop => prop.Name)
-                    .ToArray();
-
-                // Monta a lista de parâmetros da consulta SQL para atualização
-                var parameters = columnNames.Select(colName => $"{colName} = @{colName}").ToArray();
-
-                // Monta a consulta SQL de atualização dinâmica
-                string query = $"UPDATE {TableName} SET {string.Join(", ", parameters)} WHERE Id = @Id";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    // Configura os parâmetros da consulta SQL com os valores atualizados do objeto 'item'
-                    foreach (var propInfo in typeof(T).GetProperties())
-                    {
-                        var paramValue = propInfo.GetValue(item);
-                        command.Parameters.AddWithValue($"@{propInfo.Name}", paramValue ?? DBNull.Value);
-                    }
-
-                    command.ExecuteNonQuery();
-                }
+                Console.WriteLine("ID inválido.");
+                return false;
             }
-            return true;
-        }catch (SqlException sqlEx)
+
+            PropertyInfo[] propriedades = tipo.GetProperties().Where(p => p.Name != "Id").ToArray();
+
+            string setClause = string.Join(", ", propriedades.Select((p, i) => $"{p.Name} = @p{i}"));
+
+            string query = $"UPDATE {TableName} SET {setClause} WHERE Id = @Id";
+
+            using (SqlCommand command = new SqlCommand(query, ConnectionManager.GetConnection()))
+            {
+                command.Parameters.AddWithValue("@Id", id);
+
+                for (int i = 0; i < propriedades.Length; i++)
+                {
+                    command.Parameters.AddWithValue($"@p{i}", propriedades[i].GetValue(item));
+                }
+
+                ConnectionManager.OpenConnection();
+
+                int count = command.ExecuteNonQuery();
+
+                ConnectionManager.CloseConnection();
+
+                return count > 0;
+            }
+        }
+        catch (SqlException sqlEx)
         {
             Console.WriteLine("Erro de banco: " + sqlEx.Message);
             return false;
-        }catch (Exception ex)
+        }
+        catch (Exception ex)
         {
             Console.WriteLine("Erro: " + ex.Message);
             return false;
@@ -175,21 +216,35 @@ public abstract class DAO<T> : IDAO<T> where T : IEntidade
     {
         try
         {
-            using (SqlConnection connection = ConnectionManager.GetConnection())
+            Type tipo = typeof(T);
+
+            if (id <= 0)
             {
-                string query = $"DELETE FROM {TableName} WHERE Id = @Id";
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@Id", id);
-                    command.ExecuteNonQuery();
-                }
+                Console.WriteLine("ID inválido.");
+                return false;
             }
-            return true;
-        }catch(SqlException sqlEx)
+
+            string query = $"DELETE FROM {TableName} WHERE Id = @Id";
+
+            using (SqlCommand command = new SqlCommand(query, ConnectionManager.GetConnection()))
+            {
+                command.Parameters.AddWithValue("@Id", id);
+
+                ConnectionManager.OpenConnection();
+
+                int count = command.ExecuteNonQuery();
+
+                ConnectionManager.CloseConnection();
+
+                return count > 0;
+            }
+        }
+        catch (SqlException sqlEx)
         {
             Console.WriteLine("Erro de banco: " + sqlEx.Message);
             return false;
-        }catch(Exception ex)
+        }
+        catch (Exception ex)
         {
             Console.WriteLine("Erro: " + ex.Message);
             return false;
